# specs/openapi.dev.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "API Reference",
    "description": "",
    "version": "0.0.1",
    "summary": ""
  },
...
  "paths": {
...
    "/v1/magickspaces": {
      "post": {
        "summary": "Create MagickSpace",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateMagickSpaceRequest"
              },
              "example": {
                "name": "Recycled Concrete Tuna",
                "description": "The automobile layout consists of a front-engine design, with transaxle-type transmissions mounted at the rear of the engine and four wheel drive",
                "project_id": "47cc766ec1aa7b4c5bdb4d6b",
                "corpus_ids": [],
                "user_ids": [],
                "type": "GROUP"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CreateMagickSpaceResponse"
                },
                "example": {
                  "success": false,
                  "message": "Nam harum sodalitas testimonium timidus delibero caritas volup comburo thermae. Reiciendis cena usus cauda stella adfectus. Ipsum tabgo victus derideo abundans.",
                  "magickspace": {
                    "id": "43",
                    "name": "Eva Borer",
                    "description": "Mollitia curriculum voluptatibus conduco cumque cubicularis supplanto vulticulus creptio sufficio. Virga ver amo. Delego sub magni barba admoneo cinis tabella architecto. Cohaero conscendo dolorem thymum cito accedo corrupti sui accendo. Magnam explicabo tenetur cruciamentum demens cura teneo thymum. Blanditiis torrens altus villa pecto thorax quas casso. Tempora totidem absens ventito peior succedo cernuus dapifer. Supellex carus teneo.",
                    "corpus_ids": [
                      "eu",
                      "Lorem ex"
                    ],
                    "user_ids": [
                      "in ex Lorem dolore adipisicing",
                      "commodo",
                      "consectetur exercitation aliquip eiusmod deserunt"
                    ],
                    "type": "group",
                    "created_at": "2025-12-07T14:27:38.210Z",
                    "updated_at": "2026-06-11"
                  }
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "get": {
        "summary": "Get MagickSpace List",
        "deprecated": false,
        "description": "Retrieve list of magickspaces for a user",
        "tags": [
          "MagickSpaces"
        ],
        "parameters": [
          {
            "name": "user_id",
            "in": "query",
            "description": "",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetMagickSpaceListResponse"
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/v1/magickspaces/{id}": {
      "get": {
        "summary": "Get MagickSpace by ID",
        "deprecated": false,
        "description": "",
        "tags": [],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "6823981bc07297c9f0abc354",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {}
                }
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "summary": "Delete MagickSpace",
        "deprecated": false,
        "description": "Delete a magickspace by ID",
        "tags": [
          "MagickSpaces"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "",
            "required": true,
            "example": "",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "headers": {}
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
...
  "components": {
    "schemas": {
...
      "MagickSpaceType": {
        "type": "string",
        "enum": [
          "PRIVATE",
          "GROUP"
        ]
      },
...
      "MagickSpaceSchema": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "corpus_ids": {
            "type": [
              "array",
              "null"
            ],
            "items": {
              "type": "string"
            }
          },
          "user_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "$ref": "#/components/schemas/MagickSpaceType"
          },
          "created_by": {
            "type": "string"
          },
          "updated_by": {
            "type": "string"
          },
          "created_at": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "name",
          "description",
          "corpus_ids",
          "user_ids",
          "type",
          "created_at",
          "updated_at",
          "created_by",
          "updated_by",
          "project_id"
        ]
      },
      "CreateMagickSpaceRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string"
          },
          "description": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "corpus_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "user_ids": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "type": {
            "$ref": "#/components/schemas/MagickSpaceType"
          }
        },
        "required": [
          "name",
          "corpus_ids",
          "user_ids",
          "type",
          "project_id"
        ]
      },
      "CreateMagickSpaceResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseSchema"
          },
          {
            "type": "object",
            "properties": {
              "magickspace": {
                "$ref": "#/components/schemas/MagickSpaceSchema"
              }
            },
            "required": [
              "magickspace"
            ]
          }
        ]
      },
...
      "GetMagickSpaceListResponse": {
        "allOf": [
          {
            "$ref": "#/components/schemas/BaseSchema"
          },
          {
            "type": "object",
            "properties": {
              "magickspaces": {
                "type": "array",
                "items": {
                  "$ref": "#/components/schemas/MagickSpaceSchema"
                }
              }
            }
          }
        ]
      },
...
      "BaseSchema": {
        "type": "object",
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string"
          }
        },
        "required": [
          "success",
          "message"
        ]
      },
...
    },
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "servers": [],
  "security": [
    {
      "bearer": []
    }
  ]
}